requirements.txt
django
djangorestframework
drf-spectacular
django-filter
djoser
djangorestframework_simplejwt
django-allauth
mysqlclient

# eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIiwiZXhwIjoxNzMyNDM4NDcwLCJpYXQiOjE3MzIzNTIwNzAsImp0aSI6IjE4YTg4NDRlZjBhYTQ1OTY5OWZjNjQ1YThmMjcyY2ZmIiwidXNlcl9pZCI6NH0.wPy-s_aCNIY-h6BUEPo47gUNrgv3ZyheuAZPalzq9ns


#     "refresh": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoicmVmcmVzaCIsImV4cCI6MTczMjQzODQ3MCwiaWF0IjoxNzMyMzUyMDcwLCJqdGkiOiI1NmUyNjU5ZjBkNTQ0MTdiYTA3MjQxMWU2YTVhYjRhNSIsInVzZXJfaWQiOjR9.jOdqan_ivObo5LmoSh0C_O7Lw4IUvLMHbjED9M_FY3E",
#     "access": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIiwiZXhwIjoxNzMyNDM4NDcwLCJpYXQiOjE3MzIzNTIwNzAsImp0aSI6IjE4YTg4NDRlZjBhYTQ1OTY5OWZjNjQ1YThmMjcyY2ZmIiwidXNlcl9pZCI6NH0.wPy-s_aCNIY-h6BUEPo47gUNrgv3ZyheuAZPalzq9ns"

